+++ src/Integration/SoapIntegration/CountryApi/Exception/CountryApiException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace App\Integration\SoapIntegration\CountryApi\Exception;
+
+class CountryApiException extends \Exception
+{
+}